pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "job-app-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.47.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "selenium (>=4.39.0,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
